Show viewer name on reflection card input
